feat(Contacts): add util component
