+++ old_code/backup_style.css
@@ -56,7 +56,7 @@
 }
 
 
-/*TODO: Unknown usage*/
+
 .text {
     max-width:100%;
     padding-top: 182px;
