+++ src/components/Contacts/index.module.css
@@ -1,0 +1,55 @@
+.contact_with_us-title {
+    margin-bottom: 40px;
+}
+
+/*TODO: найти похожие конструкции background: url и переделать на img*/
+.contact-img {
+    background: url("./img/icon/phone_book-icon.png");
+}
+
+.more-img {
+    background: url("./img/icon/flower-icon.png");
+    width: 19px;
+    height: 19px;
+}
+
+.container-btn {
+    margin-left: 32px;
+    padding-top: 6px;
+}
+
+.contact-btns {
+    margin-top: 40px;
+}
+
+.img-right-contact {
+    background: url("./img/assets/articles-bg.jpg") no-repeat center center/contain;
+    height: 410px;
+    margin-top: 88px;
+}
+
+.desk {
+    display: none;
+}
+
+.btn {
+    background: #24aa17 !important;
+    height: 50px;
+    width: 200px;
+    color: white;
+    border-radius: 0px;
+
+}
+
+.text {
+    max-width: 100%;
+    padding-top: 182px;
+    padding-bottom: 190px;
+}
+
+@media screen and (min-device-width: 768px) {
+    .main-container {
+        max-width: 1260px;
+        margin: 0 auto;
+    }
+}
